% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_functions.R
\name{apply_limits}
\alias{apply_limits}
\title{Limit output to between 0 and 100 for data which is a percentage}
\usage{
apply_limits(forecast)
}
\arguments{
\item{forecast}{a forecast object}
}
\description{
ensures that predicted values are between 0 and 100
}
