% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{do_forecasts}
\alias{do_forecasts}
\title{Compare a number of forecasting methods on one time series}
\usage{
do_forecasts(all_forecasting_functions, time_series, forecasting_horizon,
  max_forecasting_horizon, location, ab, organism, frequency,
  apply_limits = TRUE)
}
\arguments{
\item{all_forecasting_functions}{list of lists describing the names of forecasting functions to be compared. Follow structure of list of same name exported from module}

\item{time_series}{output of convert_to_raw_time_series}

\item{location}{where was data taken from for plot title}

\item{ab}{name of antibiotic for plot title}

\item{organism}{name of organism for plot title}

\item{frequency}{12 for months, 4 for quarters}

\item{apply_limits}{true or false, limit forecasts to 0-100}
}
\description{
passes time series produced by convert_to_raw_time_series to different forecasting algorithms and then collates the results
}
