% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{inspect_residuals}
\alias{inspect_residuals}
\title{Inspects Residuals of Model fit}
\usage{
inspect_residuals(fit, residuals, forecast_name)
}
\arguments{
\item{fit}{a fitted forecsting model}

\item{residuals}{the residuals of the fitted model}

\item{forecast_name}{name of the forecast object}
}
\description{
returns the mean of the residuals and a p value of pormanteau test for normallity
}
