% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{ME}
\alias{ME}
\title{Mean error of forecast}
\usage{
ME(testing_data, point_forecasts)
}
\arguments{
\item{testing_data}{the end of the time series that was removed for forecasting}

\item{point_forecasts}{mean from a forecast object}
}
\description{
calculates mean error of a forecast
}
