% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{my_autolayer}
\alias{my_autolayer}
\title{My adaption of autolayer function}
\usage{
my_autolayer(object, series = NULL, ...)
}
\arguments{
\item{object}{a mean of a forecast object}

\item{series}{name of series to be plotted}
}
\description{
adapted from forecast:::autolayer to allow objects of length 1 to be plotted also
}
