% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{my_checkresiduals}
\alias{my_checkresiduals}
\title{My adaptation of check_residuals function}
\usage{
my_checkresiduals(object, lag, df = NULL, test, plot = TRUE, ...)
}
\arguments{
\item{object}{a fitted forecasting model}
}
\description{
adapted from forecast::checkresiduals
}
