% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_functions.R
\name{nn_autoregression_forecast}
\alias{nn_autoregression_forecast}
\title{Neural Network Autoregression Forecast}
\usage{
nn_autoregression_forecast(training_data, forecasting_horizon, frequency,
  apply_limits = TRUE)
}
\arguments{
\item{training_data}{time series to be forecasted}

\item{forecasting_horizon}{time in units of training_data to be forecasted}

\item{frequency}{12 for months, 4 for quarters etc}

\item{apply_limits}{keep data with 0 and 100, default is TRUE}
}
\description{
fits an autoregressive feed forward neural network model to the data and estimates uncertainity with bootstrapping
}
