% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{PI_width}
\alias{PI_width}
\title{Prediction Interval Width}
\usage{
PI_width(lower, upper)
}
\arguments{
\item{lower}{a lower prediction interval}

\item{upper}{an upper prediction interval}
}
\description{
calculates the width of the forecast prediction intervals
}
