% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{reverse_differencing}
\alias{reverse_differencing}
\title{Reverse differencing}
\usage{
reverse_differencing(time_series, differencing_order)
}
\arguments{
\item{time_series}{any time series object}

\item{differencing_order}{order of differencing that was applied}
}
\description{
reverts differencing that was applied to make training data stationary
}
