% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Squared Error}
\usage{
RMSE(testing_data, point_forecasts)
}
\arguments{
\item{testing_data}{the end of the time series that was removed for forecasting}

\item{point_forecasts}{mean from a forecast object}
}
\description{
calculates root mean squared error of a forecast
}
