% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_comparisons.R
\name{SMAPE}
\alias{SMAPE}
\title{Symmetirc mean absolute percentage error}
\usage{
SMAPE(testing_data, point_forecasts)
}
\arguments{
\item{testing_data}{the end of the time series that was removed for forecasting}

\item{point_forecasts}{mean from a forecast object}
}
\description{
calculates the symmetric mean absolute percentage error
}
