% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_functions.R
\name{t_and_l_forecast}
\alias{t_and_l_forecast}
\title{Teodoro and Lovis Forecast}
\usage{
t_and_l_forecast(training_data, forecasting_horizon, frequency,
  apply_limits = TRUE, filter = "DECF")
}
\arguments{
\item{training_data}{time series to be forecasted}

\item{forecasting_horizon}{time in units of training_data to be forecasted}

\item{frequency}{12 for months, 4 for quarters etc}

\item{apply_limits}{keep data with 0 and 100, default is TRUE}

\item{filter}{filter to be used for selecting components of EMD for KNN forecasting. Default is DECF}
}
\description{
fits the model developed by Teodoro and Lovis, 2013
}
