% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Writes the model output to file}
\usage{
write_output(output_list, output_file)
}
\arguments{
\item{output_list}{list of forecast summaries produced by do_forecasts()}

\item{output_file}{name of output file}
}
\description{
takes a list of outputs from do_forecasts() which summarise all forecasting results and writes them to file
}
